src/django_app/tests/conftest.py
from pathlib import Path
import pytest
from rest_framework.test import APIClient


@pytest.fixture
def resources_path():
    return Path("./tests/resources/").resolve()

@pytest.fixture
def tmp_path():
    return Path("./tests/tmp/").resolve()

@pytest.fixture
def api_client() -> APIClient:
    return APIClient()

@pytest.fixture
def mock_telegram_service(mocker):
    return mocker.patch("tables.services.telegram_trigger_service.TelegramTriggerService.register_telegram_trigger")